Moving the src/app/services/*.spec.ts
