+++ sonar-project.properties
@@ -3,7 +3,7 @@ sonar.projectName=rpx-xui-mo-webapp
 sonar.projectVersion=0.0.1
 
 sonar.sources=src/app/
-sonar.exclusions= api/**, test/**, config/**, api/**/templates/**, src/public/**, src/environments/**, **/*.html, **/*.scss, src/app/app.constants.ts
+sonar.exclusions= api/**, test/**, config/**, api/**/templates/**, src/public/**, src/environments/**, **/*.html, **/*.scss, src/app/app.constants.ts, src/app/services/dummyAppInsights.ts
 sonar.test.inclusions=src/app/**/*.spec.ts, src/app/*.spec.ts, src/app/services/*.spec.ts
 sonar.dynamicAnalysis=reuseReports
